requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pdfplumber==0.11.4
pandas==2.2.3
openpyxl==3.1.5
supabase==2.11.0
boto3==1.35.0
python-dotenv==1.0.1
PyYAML==6.0.2
tenacity==9.0.0
schedule==1.2.2
pytest==8.3.5
